requirements.txt
configspace
python >= 3.7
pytorch == 1.9
pillow
simplejson
scikit-learn
tensorboard
yacs